Fix casting error (#459)
Fix casting error

+++ tool/yubico-piv-tool.c
@@ -1561,7 +1561,7 @@ static void print_cert_info(ykpiv_state *state, enum enum_slot slot, const EVP_M
   fprintf(output, "Slot %x:\t", slot_name);
 
   unsigned char certdata[YKPIV_OBJ_MAX_SIZE * 10] = {0};
-  unsigned long certdata_len = sizeof(certdata);
+  size_t certdata_len = sizeof(certdata);
   if(ykpiv_util_get_certdata(data, len, certdata, &certdata_len) != YKPIV_OK) {
     fprintf(output, "Failed to get certificate data\n");
     return;
